server/requirements.txt
python-dotenv
flask
flask-cors
pymongo
